tests/requirements.txt
pytest
pytest-cov
pytest-asyncio
httpx
fastapi
sqlalchemy
uvicorn
pydantic-settings
pydantic>=2.0.0,<3.0.0
meilisearch
python-tds
sqlalchemy-pytds
email-validator
passlib
PyJWT
# delete
rdflib